+++ src/widgets/CommonDetailComponent.tsx
